Texture and Buffer inputs now also have stable names per shader. Some code cleanup. #274
